thuc tap 29/9

+++ thuctap/src/main/java/store/polyfood/thuctap/services/CartItemService.java
@@ -5,12 +5,15 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import store.polyfood.thuctap.models.entities.*;
 import store.polyfood.thuctap.models.responobject.Response;
 import store.polyfood.thuctap.repositories.CartItemRepo;
 import store.polyfood.thuctap.repositories.CartRepo;
 import store.polyfood.thuctap.repositories.ProductRepo;
+import store.polyfood.thuctap.repositories.UserRepo;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
@@ -19,6 +22,10 @@ import java.util.Map;
 @Service
 public class CartItemService implements ICartItemService {
 
+    @Autowired
+    private AccountService accountService;
+    @Autowired
+    private UserRepo userRepo;
     @Autowired
     private CartItemRepo cartItemRepo;
     @Autowired
@@ -29,10 +36,16 @@ public class CartItemService implements ICartItemService {
 
     @Override
     public Response createNew(CartItem request) {
-        Carts carts = cartRepo.findById(request.getCartId()).orElse(null);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username =  (String) authentication.getPrincipal();
+        Account account = (Account) accountService.loadUserByUsername(username);
+        User user = userRepo.findByAccountId(account.getAcountId());
+        Carts carts = cartRepo.findAllByUserId(user.getUserId());
         if (carts == null) {
-            return  new Response<>(LocalDateTime.now().toString(),
-                    404, "Cart not found", null);
+            Carts carts1 = new Carts();
+            carts1.setUser(user);
+            cartRepo.save(carts1);
+            carts = carts1;
         }
         Product product = productRepo.findById(request.getProductId()).orElse(null);
         if (product == null) {
@@ -44,7 +57,6 @@ public class CartItemService implements ICartItemService {
         request.setCreatedAt(LocalDateTime.now());
         cartItemRepo.save(request);
         return new Response<>(LocalDateTime.now().toString(), 200, null ,"Success");
-
     }
 
     @Override
@@ -62,20 +74,24 @@ public class CartItemService implements ICartItemService {
 
     @Override
     public Response update(CartItem request) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username =  (String) authentication.getPrincipal();
+        Account account = (Account) accountService.loadUserByUsername(username);
+        User user = userRepo.findByAccountId(account.getAcountId());
         CartItem cartItem = cartItemRepo.findById(request.getCartItemId()).orElse(null);
         if (cartItem == null) {
             return  new Response<>(LocalDateTime.now().toString(),
                     404, "Cart item not found", null);
         }
+        Carts carts  = cartItem.getCarts();
+        if(!user.getCarts().contains(carts)) {
+            return  new Response<>(LocalDateTime.now().toString(),
+                    403, "Forbidden", null);
+        }
         Product product = productRepo.findById(request.getProductId()).orElse(null);
         if (product == null) {
             return  new Response<>(LocalDateTime.now().toString(),
                     404, "Product not found", null);
-        }
-        Carts carts = cartRepo.findById(request.getCartId()).orElse(null);
-        if (carts == null) {
-            return  new Response<>(LocalDateTime.now().toString(),
-                    404, "Cart not found", null);
         }
         request.setProduct(product);
         request.setCarts(carts);
@@ -88,10 +104,19 @@ public class CartItemService implements ICartItemService {
 
     @Override
     public Response delete(int id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username =  (String) authentication.getPrincipal();
+        Account account = (Account) accountService.loadUserByUsername(username);
+        User user = userRepo.findByAccountId(account.getAcountId());
         CartItem cartItem = cartItemRepo.findById(id).orElse(null);
         if (cartItem == null) {
             return  new Response<>(LocalDateTime.now().toString(),
                     404, "Cart item not found", null);
+        }
+        Carts carts  = cartItem.getCarts();
+        if(!user.getCarts().contains(carts)) {
+            return  new Response<>(LocalDateTime.now().toString(),
+                    403, "Forbidden", null);
         }
         cartItemRepo.delete(cartItem);
         return new Response(LocalDateTime.now().toString(),
